requirements.txt
crewai==1.0.0
crewai_tools==0.1.0
fastapi==0.109.0
google-generativeai==0.3.2
langchain==0.1.0
langchain-community==0.0.13
langchain-google-genai==0.0.5
pydantic==2.5.3
python-dotenv==1.0.0
uvicorn==0.25.0 